Update some info
Just updated some of the most outdated info. Not a permanent fix, but i think like this it should be fine for now. Once ubuntu-touch.io is live and unity8 development has really taken off, we can get some people with actual writing skills on this job and re-build the website with the same technology ubuntu-touch.io uses. But for now, this should do it. It's just a dev website after all.